refactor: 파일 분리 및 클로드 리뷰 반영
